Split tests for VersionedEntriesTest
